Player cuztomization extended and floor tile bug fix All weapon attirbutes can now be customized through the PlayerCustomizationUI. Fixed a case where a enemy standing outside or close tothe border to a territory crashed the game.

+++ All Them Pixels/All Them Pixels/Player.cpp
@@ -178,7 +178,7 @@ void Player::update(UpdateInfo info)
 	Vector2f spawn = aimBoxPosition;
 	Vector2f direction = Vector2fMath::unitVector(getJoystickVector(Joystick::Axis::U, Joystick::Axis::R));
 	
-	speed = getJoystickVector(Joystick::Axis::X, Joystick::Axis::Y) * config.speed;
+	speed = getJoystickVector(Joystick::Axis::X, Joystick::Axis::Y) * (float)config.speed;
 	aimVector = direction;
 
 	updateAim();
